Create ThirdProject.cpp
Use the open space in review.cpp to practice what you learned about variables. Here are some ideas:

Create a program that takes in the weight of an item and then calculates how much that item would weigh on Mars.
Create a program that asks for a distance in miles as input. The program will then output how much that distance is in kilometers.
(You can also just run the code as is to move on.)